refactor: consolidate auto-form utility functions and update TypeScript configuration
- Merged utility functions from `utils` to `helpers` for better organization.
- Updated TypeScript paths in `tsconfig.json` for simplified imports.
- Introduced new shared types for input properties in `shared-form-types.ts`.
- Added a new `helpers.tsx` file containing various utility functions for form handling.
- Adjusted imports across auto-form components to reflect the new structure.
