+++ components/ui/auto-form/fields/date.tsx
@@ -3,7 +3,7 @@ import { FormControl, FormItem, FormMessage } from "@/components/ui/form";
 import AutoFormLabel from "../common/label";
 import AutoFormTooltip from "../common/tooltip";
 import type { AutoFormInputComponentProps } from "../types";
-import { getBaseType } from "../utils";
+import { getBaseType } from "../helpers";
 
 /**
  * Convert a value to a Date object if needed.
