+++ components/ui/auto-form/fields/radio-group.tsx
@@ -9,7 +9,7 @@ import * as z from "zod";
 import AutoFormLabel from "../common/label";
 import AutoFormTooltip from "../common/tooltip";
 import type { AutoFormInputComponentProps } from "../types";
-import { getBaseSchema } from "../utils";
+import { getBaseSchema } from "../helpers";
 
 /**
  * Get enum values from a ZodEnum schema.
